Move hasFrameDebugInfo, getRegValueAtFrame into Object-elf.C
I have no idea why these were in a separate file. Object-elf.C assumes
DWARF facilities are present.
